+++ lib/presentation/stage_create/bloc/screen/stage_create_state.dart
@@ -6,4 +6,8 @@ class CreatingStageCreate extends StageCreateState {}
 
 class SuccessStageCreate extends StageCreateState {}
 
-class ErrorStageCreate extends StageCreateState {}
+class ErrorStageCreate extends StageCreateState {
+  final String errorMessage;
+
+  ErrorStageCreate(this.errorMessage);
+}
